add tests for generated logs
